==> Lessons/06_Lab_GLMs.Rmd <==
---
title: "6: Lab - Generalized Linear Models"
author: "Environmental Data Analytics | John Fay and Luana Lima | Developed by Kateri Salk"
date: "Spring 2023"
output: pdf_document
geometry: margin=2.54cm
editor_options: 
  chunk_output_type: console
---

## Objectives
1. Answer questions on M5/A5
2. Answer questions on M6 - GLMs
3. Practice more application GLM to real datasets


## Set up
```{r, message = FALSE}
library(tidyverse)
library(agricolae)   
library(here)
here()

EPAair <- read.csv(here("Data/Processed_KEY/EPAair_O3_PM25_NC1819_Processed.csv"), stringsAsFactors = TRUE)
# Set date to date format
EPAair$Date <- as.Date(EPAair$Date, format = "%Y-%m-%d")

Litter <- read.csv(here("Data/Processed_KEY/NEON_NIWO_Litter_mass_trap_Processed.csv"), stringsAsFactors = TRUE)
# Set date to date format
Litter$collectDate <- as.Date(Litter$collectDate , format = "%Y-%m-%d")

# Set theme
mytheme <- theme_classic(base_size = 14) +
  theme(axis.text = element_text(color = "black"), 
        legend.position = "top")
theme_set(mytheme)
```



## Visualization and interpretation challenge

Create three plots, each with appropriately formatted axes and legends. Choose a non-default color palette.

1. geom_density of ozone divided by year (distinguish between years by adding transparency to the geom_density layer).
2. geom_boxplot of ozone divided by year. Add letters representing a significant difference between 2018 and 2019 (hint: stat_summary). 
3. geom_violin of ozone divided by year, with the 0.5 quantile marked as a horizontal line. Add letters representing a significant difference between 2018 and 2019. 

```{r, fig.height=3, fig.width=4, warning=TRUE}
#Exercise 1:
ozone.density<- ggplot(EPAair, aes(x=Ozone, fill= as.factor(Year)))+
  geom_density(alpha=0.5)+#alpha changes transparency
  scale_fill_manual(values = c("darkgreen","darkblue"))+
  labs(fill="",x="Ozone AQI values", y="Density")#adding empty "" means you add nothing

print(ozone.density)

ozone.box<- ggplot(EPAair,aes(x=as.factor(Year), y= Ozone))+
            geom_boxplot()+
            stat_summary(geom="text", fun=max, vjust=-1, size=4, label=c("a","b"))
print(ozone.box)

ozone.violin<- ggplot(EPAair,aes(x=as.factor(Year), y= Ozone))+
            geom_violin(draw_quantiles = 0.5)+
            stat_summary(geom="text", fun=max, vjust=-1, size=4, label=c("a","b"))+
            labs(x= "", y= "Ozone AQI values")+
            ylim(0,150)
print(ozone.violin)

```

## Linear Regression

Important components of the linear regression are the correlation and the R-squared value. The **correlation** is a number between -1 and 1, describing the relationship between the variables. Correlations close to -1 represent strong negative correlations, correlations close to zero represent weak correlations, and correlations close to 1 represent strong positive correlations. The **R-squared value** is the correlation squared, becoming a number between 0 and 1. The R-squared value describes the percent of variance accounted for by the explanatory variables. 

For the NTL-LTER dataset, can we predict PM2.5 from Ozone?

```{r}

#Exercise 2: Run a linear regression PM2.5 by Ozone. Find the p-value and R-squared value. 
ex_02<- lm(data = EPAair, PM2.5~Ozone)#y~x is y (dependent variable) is a function of x (independent variable)
summary(ex_02) 
# lm runs a linear regression, does not spit out a plot



#Exercise 3: Build a scatterplot. Add a line and standard error for the linear regression. Add the regression equation to the plot
ggplot(EPAair, aes(x=Ozone, y= PM2.5))


```


## AIC to select variables

What other variables can we add to improve model?

```{r}

#Exercise 4: Build correlation plots and identify more possible explanatory variables to add to the regression.
ozone.sitename<- EPAair%>%
  select(Month,PM2.5, Ozone)%>% #needed to use only numerical variables, you cannot use site name
  na.omit() #only omits na in a specific column drop.na will omit na's for all columns

ozone.corr <- cor(ozone.sitename)
corrplot(ozone.corr, method = "ellipse")#different plot than mixed
corrplot.mixed(ozone.corr, upper = "ellipse") #puts the labels inside the graph
  



#Exercise 5: Choose a model by AIC in a Stepwise Algorithm. Do the results from AIC match the variables you selected on Exercise 4?
ozonebyall.regression<- lm(data= ozone.sitename, PM2.5~Ozone+ Year+Month+ mean_lng)



#Exercise 6: Run another regression using the variables selected on Exercise 6. Compare r-squared value with the one from Exercise 2. 

```



## Litter Exercise

```{r}
# Wrangle the data
Litter.Totals <- Litter %>%
  group_by(plotID, collectDate, nlcdClass) %>%
  summarise(dryMass = sum(dryMass))

# Format ANOVA as aov
Litter.Totals.anova <- aov(data = Litter.Totals, dryMass ~ plotID)
summary(Litter.Totals.anova)

# Extract groupings for pairwise relationships
Litter.Totals.groups <- HSD.test(Litter.Totals.anova, "plotID", group = TRUE)
Litter.Totals.groups$groups

Litter.Totals <- Litter.Totals %>%
  mutate( treatgroups = Litter.Totals.groups$groups[plotID,2])

# Graph the results
Litter.Totals.plot <- ggplot(Litter.Totals, aes(x = plotID, y = dryMass)) +
  geom_boxplot() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  stat_summary(geom = "text", fun = max, vjust = -1, size = 3.5,
               label = c("abc", "ab", "abcd", "bcd", "bcd", "a", 
                         "d", "abcd", "cd", "d", "abcd", "abcd")) +
  labs(x = "Plot ID", y = "Dry Mass (g)") +
  ylim(0, 85)
print(Litter.Totals.plot)

```

```{r}
#Exercise 7: Improve the plot


```
